throw exception when setting alias incorrectly
This is currently a limitation of the code. The offending lines:

```
$sbarr = range(0, 32);

if ($this->aliasForTotalNumberOfPages !== null) {
    $sbarr = range(0, 57);
}
```

If the font is loaded with range(0, 32), the alias does not work later
on. Maybe I can refactor this later so the limitation is gone, but for
now, it's there (as it is in the original code, although undocumented).